Enhancement: refactor form field collection logic and introduce collectFormFieldExposedValues utility
- Updated the Form component to utilize the new collectFormFieldExposedValues function for improved field collection, allowing for better handling of nested containers.
- Introduced OPAQUE_FORM_CONTAINERS in FormUtils to define which container types should be excluded from field collection.
- Enhanced comments for clarity on the purpose and behavior of the new logic, ensuring maintainability and understanding of the field collection process.
